Feat: Collection Component 분리
- Collection, Collections component 생성 및 분리
- 파일 이름 변경(list 에서 collection으로 변경)
