+++ src/components/myPage/index.js
@@ -1,3 +1,4 @@
-export { default as ListButton } from './ListButton';
-export { default as MyList } from './MyList';
+export { default as CollectionButton } from './CollectionButton';
+export { default as Collection } from './Collection';
+export { default as Collections } from './Collections';
 export { default as SubscribeBadges } from './SubscribeBadges';
